fix(rn-offline-expo): bridge trackRef as Double, cb as Long (EAN-13)
Follow-up to #1: the Expo Android wrapper still passed Int to the new
TrackRef(cb: Long) signature (compile error), and Map<String, Int>
would truncate EAN-13 barcodes at the JS bridge regardless. Bridge
trackRef maps as Double (exact for cb < 2^53) and convert to Long/Int.
iOS unaffected (Swift Int is 64-bit).

Synced from upstream via sync-demos.sh.

+++ demos/react-native/modules/offline/android/src/main/java/cc/musicme/offline/expo/OfflineNativePlayer.kt
@@ -55,7 +55,7 @@ class NativePlayer(context: Context, appContext: AppContext) : ExpoView(context,
         }
     }
 
-    fun load(cb: Int, disc: Int, track: Int) {
+    fun load(cb: Long, disc: Int, track: Int) {
         val svc = OfflineSingleton.service ?: run {
             onError(mapOf("message" to "player_not_configured"))
             return
